apps/sdl/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "usf-sdl"
version = "0.1.0"
description = "USF SDL compiler service — YAML→OWL+SQL+R2RML"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlmodel>=0.0.19",
    "psycopg[async]>=3.2",
    "loguru>=0.7",
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "rdflib>=7.0",
    "owlrl>=6.0",
    "sqlglot>=25.0",
    "pyyaml>=6.0",
    "usf-rdf @ file:///${PROJECT_ROOT}/../../packages/rdf",
    "usf-core @ file:///${PROJECT_ROOT}/../../packages/core",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.hatch.build.targets.wheel]
packages = ["usf_sdl_svc"]
